feat: release new calculator
