+++ pages/enkel-kalkulator.tsx
@@ -1,0 +1,1 @@
+export { default } from 'src/salary-calculator';
